Rework local IO managers to use new input types as discussed
In particular:
 - TopLevelMetadata assets will return the actual classes instead of
   dataframes as this avoids unnecessary conversions between
   classes/dataframes (a common operation is extracting a field, e.g. an
   ID, in a downstream asset and while that can be done with
   df["field_name"].iloc[0] it's clenaer to just use class.field_name)
 - Geometry assets will return a list of (GeometryMetadata, gdf,
   names_df) because they cannot be separately partitioned

The new IO managers have been disabled in this commit because I have
introduced dependencies between some of the outputs. For example, the
SourceDataRelease requires the ID of the GeometryMetadata which is
constructed at runtime by the geometry asset.

Since our IO managers only allow for output and not input, it is
impossible to produce any assets that depend on those using the new IO
managers.

The way around this is to use the default IO managers for these assets
themselves, and then create NEW assets which depend on these assets and
exist solely to handle the output to local / Azure. Indeed, these 'new'
assets can just be cloud sensors (which leads us VERY nicely into that
bit of work which we were always going to do!)

NOTE: This commit will break the Azure IO managers. They will be fixed
soon.