Add fix_types() method to ensure parquet types are correct
This commit implements a class method called fix_types, which is called
whenever a list of metadata classes is serialised to a dataframe.
cls.fix_types(df) returns a new df where the types of columns are
properly coerced to what they should be in the resulting parquet file.
This avoids issues with pandas automatically inferring e.g. a None type
for a string column that just happens to all be Nones, and ensures that
dataframes from different countries can be concatenated by the CLI.

Closes #106